enhancement: add base class for linked lists (#6)
Create a base class for linked lists, with an abstract method _link defined in the derived classes SinglyLinkedList and DoublyLinkedList
